requirements.txt
Mastodon.py
python-dotenv
hdfs
pandas
happybase